Move helper tests to one file and add missing coverage
